% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduleWeatherForecastDownload.R
\name{deleteWeatherForecastDownload}
\alias{deleteWeatherForecastDownload}
\title{Delete Scheduled Weather Forecast Download}
\usage{
deleteWeatherForecastDownload(taskname)
}
\arguments{
\item{taskname}{Name of the task for the Windows Task Scheduler}
}
\description{
Function deleting a scheduled automatic retreival of weather forecast from Weather Underground
}
\seealso{
\link[GermanWeatherData]{scheduleWeatherForecastDownload}, \link[taskscheduleR]{taskscheduler_delete}
}

