% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCode.R
\name{geoCode}
\alias{geoCode}
\title{GEO Code}
\usage{
geoCode <- function(address, ...)
}
\arguments{
\item{address}{Character value containing the address in te form "Street Nr., ZIP, City, Country"}

\item{proxy_opts}{Optional list object containing proxy inormation required for online connection}

\item{google_maps_api}{String containing the root of Google Maps' API}

\item{return.call}{String parameter to the API, defining the return type (default = "json")}

\item{verbose}{Optional parameter triggeringa console notification of the current query (if set to TRUE)}
}
\value{
Returns a simle vector containing two elements - latitude and longitude
}
\description{
Function retreiving the GEO coordinates of an address, utilizing the Google Maps API
}
\seealso{
\link[GermanWeatherData]{getWeatherData},  \link[GermanWeatherData]{getWeatherForecast}
}

