% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeatherData.R
\name{getWeatherData}
\alias{getWeatherData}
\title{Get Historical Weather Data}
\usage{
getWeatherData(address, timeseries_start_date = NULL, timeseries_end_date = NULL, weather_data = c("temperature", "humidity", "sunshine", "wind", "precipitation"),
proxy_url, proxy_port, proxy_password = NULL, proxy_username, verbose = FALSE)
}
\arguments{
\item{address}{Character value or vector of character strings (for query of multiple locations) containing the address of the location for which the data should be retreived in the format "street name, street number, ZIP code, city, country"}

\item{timeseries_start_date}{Optional start date for the weather data time series in the format "%Y-%m-%d"}

\item{timeseries_end_date}{Optional end date for the weather data time series in the format "%Y-%m-%d"}

\item{weather_data}{Optional vector of character values controlling which data elements are required (impacts the choice of weather station); one of c("temperature", "humidity", "sunshine", "wind", "precipitation")}

\item{proxy_url}{Optional character argument for the proxy URL}

\item{proxy_port}{Optional numeric argument for the proxy port}

\item{proxy_password}{Optional character argument for the password (if not supplied, no proxy settings are set)}

\item{proxy_username}{Optional character argument containing the username}

\item{verbose}{Optional parameter triggeringa console notification of the current query (if set to TRUE)}
}
\value{
Returns a list object containing data frames with historical weather data for each provided location
}
\description{
Function retreiving historical weather data based on a provided address from the server of the 'Deutscher Wetterdienst'
}
\seealso{
\link[GermanWeatherData]{retrieve_dwd_stations}, \link[GermanWeatherData]{retrieve_dwd_data}
}

