% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeatherForecast.R
\name{getWeatherForecast}
\alias{getWeatherForecast}
\title{Download weather forecast}
\usage{
getWeatherForecast(address, api_key = "a98e65d6eb8f07eb",
  max_requests = 500, export_query = FALSE,
  destination_path = file.path(Sys.getenv("USERPROFILE"), "Desktop", fsep =
  "\\\\"), proxy_url = "proxy.whu.edu", proxy_port = 3128,
  proxy_password = NULL, proxy_username = paste(Sys.getenv("USERDOMAIN"),
  Sys.getenv("USERNAME"), sep = "\\\\"), verbose = FALSE, ...)
}
\arguments{
\item{address}{Character value or vector of character strings (for query of multiple locations) containing the address of the location for which the data should be retreived in the format "street name, street number, ZIP code, city, country"}

\item{api_key}{Weather Underground API key as character}

\item{max_requests}{Numeric value indicating the maximum number of daily requests (in free version of the WU API == 500)}

\item{export_query}{Logical value indicating if the downloaded data should be exported to a .csv file}

\item{destination_path}{Destination path of the export .csv-file (default is set to desktop)}

\item{proxy_url}{Optional character argument for the proxy URL}

\item{proxy_port}{Optional numeric argument for the proxy port}

\item{proxy_password}{Optional character argument for the password (if not supplied, no proxy settings are set)}

\item{proxy_username}{Optional character argument containing the username}

\item{verbose}{Optional parameter triggeringa console notification of the current query (if set to TRUE)}

\item{...}{List of optional parameters to underlying functions}
}
\description{
Function for getting a weather forecast from an open weather API. Powered by Weather Underground (wunderground.com).
}
\seealso{
\url{https://www.wunderground.com/weather/api/d/docs}
}

