% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestWeatherStation.R
\name{linear_coordinates_distance}
\alias{linear_coordinates_distance}
\title{Calculate Euclidean Distance of Two Locations}
\usage{
linear_coordinates_distance(from_latitude, from_longitude, to_latitude, to_longitude)
}
\arguments{
\item{from_latitude}{Latitude of the origin (provided as a numeric value)}

\item{from_longitude}{Longitude of the origin (provided as a numeric value)}

\item{to_latitude}{Latitude of the destination (provided as a numeric value)}

\item{to_longitude}{Longitude of the destination (provided as a numeric value)}
}
\value{
Returns the distance of the two locations in km (as a numeric value)
}
\description{
Function calculating linear distance between two GPS coordinates expressed as decimal numbers
}
\seealso{
\link[GermanWeatherData]{geoCode}, \link[GermanWeatherData]{find_nearest_weather_station}
}

