% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveDWDstations.R
\name{retrieve_dwd_stations}
\alias{retrieve_dwd_stations}
\title{Get list of DWD weather stations}
\usage{
retrieve_dwd_stations(recording_start = NULL, recording_end = NULL, station_type = c("FF", "KL", "RR", "SO", "TU"), proxy_opts = NULL, ...)
}
\arguments{
\item{recording_start}{Optional date parameter determining when data collection must have started, to filter out irrelevant weather stations}

\item{recording_end}{Optional date parameter determining until when data collection must have been performed, to filter out irrelevant weather stations}

\item{station_type}{Optional vector of string values, defining which data types should be recorded at the weather station (used for filtering the list)}

\item{proxy_opts}{Optional list object containing proxy settings}

\item{verbose}{Optional parameter triggeringa console notification of the current query (if set to TRUE)}
}
\value{
Returns a list object containing data frames with historical weather data for each provided location
}
\description{
Function retreiving and filtering a list of weather stations from the 'Deutscher Wetterdienst' server
}
\seealso{
\link[GermanWeatherData]{getWeatherData}, \link[GermanWeatherData]{find_nearest_weather_station}
}

