% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduleWeatherForecastDownload.R
\name{scheduleWeatherForecastDownload}
\alias{scheduleWeatherForecastDownload}
\title{Schedule Weather Forecast Download}
\usage{
scheduleWeatherForecastDownload(taskname, frequency = c("MINUTE", "DAILY",
  "WEEKLY", "ONCE"), starttime = "09:00", weekdays = c("MON", "TUE", "WED",
  "THU", "FRI", "SAT", "SUN"), minutes_step = 1, address,
  api_key = "a98e65d6eb8f07eb", max_requests = 500,
  destination_path = file.path(Sys.getenv("USERPROFILE"), "Desktop", fsep =
  "\\\\"), proxy_url = "proxy.whu.edu", proxy_port = 3128,
  proxy_password = NULL, proxy_username = paste(Sys.getenv("USERDOMAIN"),
  Sys.getenv("USERNAME"), sep = "\\\\"))
}
\arguments{
\item{taskname}{Name of the task for the Windows Task Scheduler}

\item{frequency}{One of c("MINUTE", "DAILY", "WEEKLY","ONCE")}

\item{starttime}{Character in the format "%H:%M"}

\item{weekdays}{One of weekdays = c("MON", "TUE", "WED", "THU", "FRI", "SAT", "SUN"), if weekly scheduling is required}

\item{minutes_step}{Numeric value indicating the frequency of the task in minutes (only if frequency == "MINUTE")}

\item{address}{Character value or vector of character strings (for query of multiple locations) containing the address of the location for which the data should be retreived in the format "street name, street number, ZIP code, city, country"}

\item{api_key}{Weather Underground API key as character}

\item{max_requests}{Numeric value indicating the maximum number of daily requests (in free version of the WU API == 500)}

\item{destination_path}{Destination path of the export .csv-file (default is set to desktop)}

\item{proxy_url}{Optional character argument for the proxy URL}

\item{proxy_port}{Optional numeric argument for the proxy port}

\item{proxy_password}{Optional character argument for the password (if not supplied, no proxy settings are set)}

\item{proxy_username}{Optional character argument containing the username}
}
\description{
Function scheduling an automatic retreival of weather forecast for a list of addresses, utilizing the Weather Underground API
}
\seealso{
\link[GermanWeatherData]{getWeatherForecast}, \url{https://www.wunderground.com/weather/api/d/docs}
}

